+++ dbus-serialbattery/battery.py
@@ -1471,7 +1471,7 @@ class Battery(ABC):
         :return: The maximum discharge current
         """
         if self.temperature_mos is None:
-            return self.max_battery_charge_current
+            return self.max_battery_discharge_current
 
         try:
             if utils.CHARGE_MODE == 2:
@@ -1502,7 +1502,7 @@ class Battery(ABC):
             file = exception_traceback.tb_frame.f_code.co_filename
             line = exception_traceback.tb_lineno
             logger.error("Non blocking exception occurred: " + f"{repr(exception_object)} of type {exception_type} in {file} line #{line}")
-            return self.max_battery_charge_current
+            return self.max_battery_discharge_current
 
     def calc_max_charge_current_from_soc(self) -> float:
         """