+++ dbus-serialbattery/dbus-serialbattery.py
@@ -379,6 +379,7 @@ def main():
         """
         from bms.daly_can import Daly_Can
         from bms.jkbms_can import Jkbms_Can
+        from bms.lltjbd_can import LltJbd_Can
         from bms.rv_c_can import RV_C_Can
         from bms.ubms_can import Ubms_Can
 
@@ -386,6 +387,7 @@ def main():
         supported_bms_types = [
             {"bms": Daly_Can},
             {"bms": Jkbms_Can},
+            {"bms": LltJbd_Can},
             {"bms": RV_C_Can},
             {"bms": Ubms_Can},
         ]