+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Added: Health check for batteries which are using the callback by @mr-manuel
 * Added: KS48100 BMS - Read SoH with https://github.com/mr-manuel/venus-os_dbus-serialbattery/pull/344 by @kopierschnitte
 * Added: Venus OS 3.7x GUIv2 support by @mr-manuel
+* Added: JBD CAN support by @dchervov
 * Changed: Daren 485 BMS - Fixed charge/discharge calculation with https://github.com/mr-manuel/venus-os_dbus-serialbattery/pull/343 by @kopierschnitte
 * Changed: Exit behavior for excluded devices to behave like Victron services by @mr-manuel
 * Changed: Fixed typo in activation instructions by @mr-manuel
@@ -33,6 +34,7 @@
 * Changed: Seplos BMS: Fix problems with unique identifier when daisy chained by @KoljaWindeler
 * Changed: Use Bluetooth MAC address as unique identifier for all Bluetooth BMS by @mr-manuel
 * Changed: Use port and address as unique identifier is now available for all serial BMS by @mr-manuel
+* Changed: Fixed discharge current limit calculations when MOSFET temperature is not available, by @dchervov
 
 ## v2.0.20250729